backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
numpy>=1.26.0
torch>=2.2.0
transformers>=4.37.0
requests>=2.31.0
